animation looks right

+++ src/aoc/Day17/Cross.pde
@@ -1,5 +1,5 @@
 class Cross extends Rock {
-  Cross(float x, float y) {
+  Cross(int x, int y) {
     super(x, y);
     name = "cross";
     col = color(72, 100, 100);
